Refactor PhoneVerificationForm to improve usability and structure
Reorganized the PhoneVerificationForm component to enhance code readability and functionality. Introduced `useHeaderConfig`, improved phone number validation, and added a more robust reCAPTCHA check. Additional updates include better error handling, UI enhancements, and separating utility functions for maintainability.
